Flesh out chefspec tests
This commit removes 2 of the 3 canned files that were put in
place via cookbook_file (dangerfile and dangerfile2) in favor
of using the 'content' attribute on a 'file' resource instead
inside recipes/tester.rb. The canned file serial.conf was kept
as-is since it is large.

We perform more extensive ChefSpec testing now via the new-ish
render_file('/some/path').with_content(/Regex-here/) helper.
When there is a .without_content(...) helper we can add tests
for that.