Catch lines missed by strict patterns
There are times where a pattern will match on first run and replace the line with the given value.  On second run the line will not match the pattern and will then be appended to the bottom of the file.  This is bad, especially if the file depends on formatting like XML.