Replace the rspec unit tests with inspec integration tests
The rspec tests were written for the heavy weight providers and did not
work when the providers were converted to custom resources. This change
deletes the old rspec tests and adds roughly equivalent inspec tests.

A chef-client handler was added for testing purposes to allow inspec tests
to check the updated vs up to date status of resources.  The line resource
has had problems with adding lines after every run.  Checking the results
after rerunning changes is important testing.

Custom inspec resources were added to make testing simpler.  Resources
were added to check the number of lines in a file and the number of
times a pattern matched the contents of a file.