Handle missing file errors better
The code raised an error for a missing target file, but did the raise inside of a recovery block.
The recovery block would open a temporary file and ensure that the temporary file would be closed and
unlinked.  Since the temporary file never was created the the close failed obscuring the cause of the
error.