WIP - Introduces a Dangerfile to the project. (#65)
* Remove name_property for name, as this is actually an error

* Remove reference to match_count from the file_ext matcher

Match_count was moved to a separate matcher

* Delete minitest files

* Introduces a Dangerfile to the project.

* Travis needs to bundle install if danger's around.

* Adds a reference link to Dangerfile docs.

* Handle missing file errors better

The code raised an error for a missing target file, but did the raise inside of a recovery block.
The recovery block would open a temporary file and ensure that the temporary file would be closed and
unlinked.  Since the temporary file never was created the the close failed obscuring the cause of the
error.

* Add tests for resource calls with missing files

The delete_lines, add_to_list, and delete_from_list resources should
raise an error if the target file has not been created. The added
tests verify that the right error and message is created.

* Version 1.0.3

Additional tests.
Fix missing file bugs.

* Handle a delete item edge case

Spaces as delimiters cause problems when deleting itermediate items in a list.
Make sure a space gets left behind when an item in a space delimited list is
deleted.

* Version 1.0.4

* Update the Dangerfile

Clean up cookstyle objections

* Fail has a nice user interface than raise. Use fail in the Dangerfile.

Exclude Dangerfile from cookstyle checking.
Add rubocop.yml.