Match lines containing question mark
The append_if_no_line resource was not correctly identifying the
presence of a line if that line had a question mark, due to the question
mark being interpreted as part of the regex. As a result the line would
be appended every chef run. This commit adds the question mark character
to the list of characters to escape in Line::Helper.escape_string.