Use Regexp.escape() and add a comment
the helper class here is already baked directly into ruby, and should
really get removed overall.

commented out a subtle thing i did which i think is the right thing to
do because it makes the resource more declarative and the user doesn't
need to thing about files-with-terminating-CRs, but i have a small
doubt about adding the CR before matching, so just documented that.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>