Non slurpy append_if_no_line
this i think is more or less what it would take to make
append_if_no_line both robust via using the file resource, and
to make it non-slurpy.  it passes tests and should do a single
pass across the target file and then use the file provider
machinery to move it into place.

i'm going to close this because i think its getting pretty hacky.

its possible we should patch the core file provider to make this
use case possible without hacking it up.

if there's legit bug reports where the slurpy behavior is shown
to impact people i'd support resurrecting and merging this to fix it.

and i'm submitting this as a PR just as documentation for my future
self for things to think about in core chef.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>