Add the filter_lines resource
The filter resource allows a proc to be supplied to edit the lines of a file
Add the before filter, used to insert lines before the first, last or every match of a pattern line.
Add the after filter, used to insert lines after the first, last or every match of a pattern line.
Add the filters option to use multiple filters for one filter_lines resource.
Add the replace filter