Add the safe option
Safe causes resources to check the lines to be inserted in the file against the regular expression used
to select lines.  If the regex matches inserted lines the resource will insert the same lines usually over
and over again making a file grow without limit. Safe can be set as a default for a filter resource or
set for each filter used in a resource.